requirements.txt
asttokens==2.4.1
attrs==23.2.0
bleach==6.1.0
blinker==1.7.0
category-encoders==2.6.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
colorama==0.4.6
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
Cython==3.0.10
dash==2.16.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
decorator==5.1.1
deprecation==2.1.0
executing==2.0.1
fastjsonschema==2.19.1
Flask==3.0.3
fonttools==4.51.0
idna==3.7
imbalanced-learn==0.12.2
importlib_metadata==7.1.0
ipython==8.23.0
ipywidgets==8.1.2
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_core==5.7.2
jupyterlab_widgets==3.0.10
kaggle==1.6.12
kaleido==0.2.1
kiwisolver==1.4.5
lightgbm==4.3.0
llvmlite==0.42.0
MarkupSafe==2.1.5
matplotlib==3.7.5
matplotlib-inline==0.1.7
nbformat==5.10.4
nest-asyncio==1.6.0
numba==0.59.1
numpy==1.26.4
orjson==3.10.1
packaging==24.0
pandas==2.1.4
parso==0.8.4
patsy==0.5.6
pillow==10.3.0
platformdirs==4.2.0
plotly==5.21.0
plotly-resampler==0.10.0
pmdarima==2.0.4
prompt-toolkit==3.0.43
psutil==5.9.8
pure-eval==0.2.2
pycaret==3.3.1
Pygments==2.17.2
pyod==1.1.3
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytz==2024.1
pywin32==306
referencing==0.34.0
requests==2.31.0
retrying==1.3.4
rpds-py==0.18.0
schemdraw==0.15
scikit-base==0.7.7
scikit-learn==1.4.2
scikit-plot==0.3.7
scipy==1.11.4
setuptools==69.5.1
six==1.16.0
sktime==0.26.0
stack-data==0.6.3
statsmodels==0.14.2
tbats==1.1.3
tenacity==8.2.3
text-unidecode==1.3
threadpoolctl==3.4.0
tqdm==4.66.2
traitlets==5.14.3
tsdownsample==0.1.3
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.0.2
widgetsnbextension==4.0.10
xxhash==3.4.1
yellowbrick==1.5
zipp==3.18.1
aiohttp==3.9.5
aiosignal==1.3.1
alembic==1.13.1
anyio==4.3.0
apache-airflow==2.9.0
apache-airflow-providers-common-io==1.3.1
apache-airflow-providers-common-sql==1.12.0
apache-airflow-providers-fab==1.0.4
apache-airflow-providers-ftp==3.8.0
apache-airflow-providers-http==4.10.1
apache-airflow-providers-imap==3.5.0
apache-airflow-providers-smtp==1.6.1
apache-airflow-providers-sqlite==3.7.1
apispec==6.6.0
argcomplete==3.3.0
asgiref==3.8.1
asttokens==2.4.1
attrs==23.2.0
Babel==2.14.0
bleach==6.1.0
blinker==1.7.0
cachelib==0.9.0
category-encoders==2.6.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
clickclick==20.10.2
cloudpickle==3.0.0
colorama==0.4.6
colorlog==4.8.0
comm==0.2.2
ConfigUpdater==3.2
connexion==2.14.2
contourpy==1.2.1
cron-descriptor==1.4.3
croniter==2.0.5
cryptography==42.0.5
cycler==0.12.1
Cython==3.0.10
dash==2.16.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
DateTime==5.5
decorator==5.1.1
Deprecated==1.2.14
deprecation==2.1.0
dill==0.3.8
dnspython==2.6.1
docutils==0.21.1
email_validator==2.1.1
executing==2.0.1
fastjsonschema==2.19.1
Flask==2.2.5
Flask-AppBuilder==4.4.1
Flask-Babel==2.0.0
Flask-Caching==2.1.0
Flask-JWT-Extended==4.6.0
Flask-Limiter==3.5.1
Flask-Login==0.6.3
Flask-Session==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.2.1
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2024.3.1
google-re2==1.1
googleapis-common-protos==1.63.0
greenlet==3.0.3
grpcio==1.62.2
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
imbalanced-learn==0.12.2
importlib-metadata==7.0.0
importlib_resources==6.4.0
inflection==0.5.1
ipython==8.23.0
ipywidgets==8.1.2
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_core==5.7.2
jupyterlab_widgets==3.0.10
kaggle==1.6.12
kaleido==0.2.1
kiwisolver==1.4.5
lazy-object-proxy==1.10.0
lightgbm==4.3.0
limits==3.11.0
linkify-it-py==2.0.3
llvmlite==0.42.0
lockfile==0.12.2
Mako==1.3.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.1
marshmallow-oneofschema==3.1.1
marshmallow-sqlalchemy==0.28.2
matplotlib==3.7.5
matplotlib-inline==0.1.7
mdit-py-plugins==0.4.0
mdurl==0.1.2
more-itertools==10.2.0
multidict==6.0.5
nbformat==5.10.4
nest-asyncio==1.6.0
numba==0.59.1
numpy==1.26.4
opentelemetry-api==1.24.0
opentelemetry-exporter-otlp==1.24.0
opentelemetry-exporter-otlp-proto-common==1.24.0
opentelemetry-exporter-otlp-proto-grpc==1.24.0
opentelemetry-exporter-otlp-proto-http==1.24.0
opentelemetry-proto==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-semantic-conventions==0.45b0
ordered-set==4.1.0
orjson==3.10.1
packaging==24.0
pandas==2.1.4
parso==0.8.4
pathspec==0.12.1
patsy==0.5.6
pendulum==3.0.0
pillow==10.3.0
platformdirs==4.2.0
plotly==5.21.0
plotly-resampler==0.10.0
pluggy==1.5.0
pmdarima==2.0.4
prison==0.2.1
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
pure-eval==0.2.2
pycaret==3.3.1
pycparser==2.22
Pygments==2.17.2
PyJWT==2.8.0
pyod==1.1.3
pyparsing==3.1.2
python-daemon==3.0.1
python-dateutil==2.9.0.post0
python-nvd3==0.15.0
python-slugify==8.0.4
pytz==2024.1
pywin32==306
PyYAML==6.0.1
referencing==0.34.0
requests==2.31.0
requests-toolbelt==1.0.0
retrying==1.3.4
rfc3339-validator==0.1.4
rich==13.7.1
rich-argparse==1.4.0
rpds-py==0.18.0
schemdraw==0.15
scikit-base==0.7.7
scikit-learn==1.4.2
scikit-plot==0.3.7
scipy==1.11.4
setproctitle==1.3.3
setuptools==69.5.1
six==1.16.0
sktime==0.26.0
sniffio==1.3.1
SQLAlchemy==1.4.52
SQLAlchemy-JSONField==1.0.2
SQLAlchemy-Utils==0.41.2
sqlparse==0.5.0
stack-data==0.6.3
statsmodels==0.14.2
tabulate==0.9.0
tbats==1.1.3
tenacity==8.2.3
termcolor==2.4.0
text-unidecode==1.3
threadpoolctl==3.4.0
time-machine==2.14.1
tqdm==4.66.2
traitlets==5.14.3
tsdownsample==0.1.3
typing_extensions==4.11.0
tzdata==2024.1
uc-micro-py==1.0.3
unicodecsv==0.14.1
universal_pathlib==0.2.2
urllib3==2.2.1
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==2.2.3
widgetsnbextension==4.0.10
wrapt==1.16.0
WTForms==3.1.2
xxhash==3.4.1
yarl==1.9.4
yellowbrick==1.5
zipp==3.18.1
zope.interface==6.3
